requirements.txt
transformers
torch
pandas
numpy<2
matplotlib
seaborn
scikit-learn
scikit-surprise
nltk